New commands and old commands behavior changes: cline -> 1. Open a new mados console executable in a new interface. 		 2. Mados console are not blocked.
lcline -> 1. Open a new mados console executable in the same interface.

open -> 1. Open programs(cmd, notepad, wordpad etc. + content from PATH environment variable) in a new interface.
		2. Mados console are not blocked.
Exceptions: 1. The executables from mados console executable folder are visible in mados command line. If such an executable are sent as a parameter, user is warned that the sent parameter is a file.

openF -> 1. Open a file with its default program in a new interface.
	2. Can run executable which did not use command line arguments. For other executables, undefined behavior(depends on how the executable treats this situation).
	3. Mados console are not blocked.

openFP -> 1. Open a file with a given program(from open command) in a new interface.
	2. Mados console are not blocked.
Exceptions: 1. The executables from mados console executable folder are visible in mados command line. If such an executable are sent as a first paremeter, user is warned that the sent parameter is a file.
2. If the program are cmd, regardless the file parameter, the command result is opening the cmd console without other operations.

run -> 1. Execute a program in the same interface.
Exceptions: 1. Block the mados console if the program opens a new interface(exception: .msi files).

runP -> 1. Execute a program in a new interface.
		2. Mados console are not blocked.
Exceptions: 1. runP can open text files, pdf, etc, without blocking the mados console.