Bug fixed: when user want to sort files in a directory based on extensions, if a file doesn't have an extension, the code doesn't work properly.
Bug description: if a file doesn't have an extension, it will be created a directory which start with a non-english alphabetic character. That's because in code I try to create a directory with the same name of the regular file, and that is not ok.

Solution description: create "noextension" directory if there are files without extension and move the files in that directory. Add additional check for moving no extension files because "noextension" are not contained in files names, and in the code, I move files in proper directory based on comparison between files names and extesions.